pyproject.toml
[project]
name = "clawsynth"
version = "0.1.0"
description = "ClawSynth - Synthetic workspace and query generation toolkit"
requires-python = ">=3.13,<3.14"
dependencies = [
    "markitdown>=0.1.1",
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "pytest>=8.4.2",
    "pypdf>=6.0.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
